Fix Highlighting
