+++ src/components/common/LiveCode.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import { useStep } from '../../hooks/useStep';
-import { CodeBlock } from '../Post/MarkDownViewer';
+import { CodeBlock } from '../Post/CodeBlock';
 
 type Props = {
   code: string;
